backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
reportlab==4.0.4
huggingface_hub==0.19.4
gunicorn==21.2.0